Crate astar

Traits

ReusableSearchProblem
ReusableSearchProblem is like a regular SearchProblem but without the start() and is_end() checks. Instead, the start and end will be provided when astar_r() is called.

SearchProblem
A SearchProblem is a description of the problem that will be solved with A*. Implementing this trait will describe the problem well enough that it can be solved without any more information. N is the type of one of the search-states and C is the type of the cost to get from one state to another.

Functions

astar
Perform an A* search on the provided search-problem.

astar_r